requirements.txt
flask
transformers
torch
Pillow
gTTS
gunicorn